Fixed a stupid warning [#90123196]

+++ Transfer/ViewControllers/LoginViewController.m
@@ -159,10 +159,9 @@ IB_DESIGNABLE
 		[alertView setConfirmButtonTitle:NSLocalizedString(@"button.title.ok", nil)];
 		[alertView show];
 		
-		[[GoogleAnalytics sharedInstance] sendAppEvent:@"OAuthTokenError" withLabel:[NSString stringWithFormat:@"code: %lu", error.code]];
+		[[GoogleAnalytics sharedInstance] sendAppEvent:@"OAuthTokenError" withLabel:[NSString stringWithFormat:@"code: %lu", (long)error.code]];
 		
 		return;
-
 	}
 	
 	dispatch_async(dispatch_get_main_queue(), ^{
